pyproject.toml
[project]
name = "costume-classifier"
version = "0.1.0"
description = "Doorstep costume classifier - Raspberry Pi edge computer vision system for Halloween"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.12.0.88",
    "openai>=1.68.0",
    "python-dotenv>=1.1.1",
    "ultralytics>=8.3.63",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]

[tool.ruff]
# Set the maximum line length to 88 (Black-compatible)
line-length = 88
target-version = "py310"

# Enable common linting rules
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Allow autofix for all enabled rules
lint.fixable = ["ALL"]

# Exclude common directories
extend-exclude = [
    ".venv",
    "venv",
    "__pycache__",
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces for indentation
indent-style = "space"
